refactor : `TableGroup` 관련 클래스들을 table 패키지로 이동
