+++ src/main/java/kitchenpos/ordertable/domain/TableGroupRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.tablegroup.domain;
+package kitchenpos.ordertable.domain;
 
+import kitchenpos.ordertable.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
