+++ src/test/java/kitchenpos/table/domain/TableGroupTest.java
@@ -1,0 +1,6 @@
+package kitchenpos.table.domain;
+
+public class TableGroupTest {
+
+
+}
